requirements-docs.txt
mkdocs
mkdocs-material
mkdocs-i18n
mkdocstrings[python]
pyyaml